# backend/src/main/java/com/health/backend/dto/LoginResponse.java
package com.health.backend.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class LoginResponse {
    private boolean success;
    private String message;
    private String token;
    private int userId;
    private String fullName;
    private String role;
}